size

+++ 0x00-hello_world/6-size.c
@@ -13,10 +13,10 @@ int main(void)
 	long long d;
 	float e;
 
-	printf("Size of a char: %lu byte(s)", (unsigned long)sizeof(a));
-	printf("Size of an int: %lu byte(s)", (unsigned long)sizeof(b));
-	printf("Size of long int: %lu byte(s)", (unsigned long)sizeof(c));
-	printf("Size of long long int: %lu byte(s)", (unsigned long)sizeof(d));
-	printf("Size of float: %lu byte(s)", (unsigned long)sizeof(e));
+	printf("Size of a char: %ld byte(s)", sizeof(a));
+	printf("Size of an int: %ld byte(s)", sizeof(b));
+	printf("Size of long int: %ld byte(s)", sizeof(c));
+	printf("Size of long long int: %ld byte(s)", sizeof(d));
+	printf("Size of float: %ld byte(s)", sizeof(e));
 	return (0);
 }
